ipn/ipnlocal: update tests to check event subscriptions
Update the tests to simplify the construction of AppConnector values now that a
store callback is no longer required.

Also fix a couple of pre-existing racy tests that were hidden by not being
concurrent in the same way production is.

Add code to check that the expected events are being delivered to the backend.
As in the AppConnector itself, these tests still rely on the RouteAdvertiser to
check correctness, and will need to be updated with a bus fixture separately.

Updates #15160
Updates #17192

Change-Id: Ia3ac67bfe80bc2b12076411078a2429d94aaccd8
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>