appc: remove the StoreRoutesFunc callback
RouteInfo is now published via the event bus and handled by the local backend,
so we do not need a callback to do that. However, we still need a flag to tell
us whether to treat the route info the connector has as "definitive" for
filtering purposes. Do this with a new hasStoredRoutes flag rather than using
the presence or absence of the callback itself as the signal.

Updates #15160
Updates #17192

Change-Id: I25626338969850b426dda0be6cc112e949fabd29
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>