appc: make the RouteAdvertiser field optional
Now that events are published for route updates and storage, the
RouteAdvertiser is no longer required. We cannot yet remove it because the
tests still depend on it to verify correctness. We will need to separately
update the test fixtures to remove that dependency, then we can remove the
field entirely.

Updates #15160
Updates #17192

Change-Id: Ifca57907f5682634d9a7409700d7ba0e1c999bc8
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>