cmd/tailscale/cli: print the machine auth URL after interrupted login
This fixes a bug introduced in a prior commit in #17382, specifically:

*   Create a tailnet with device approval
*   Run `tailscale up`
*   Log in with the authentication URL, but don't approve the device

Next time you run `tailscale up`, the command would hang forever, and
never prompt you to complete the machine approval.

This occurs because if you just run `tailscale up`, you enter the
`simpleUp` case which never calls `Start()`, and then you never get an
IPN notification that you're in state `NeedsMachineAuth`. Changing the
definition of `simpleUp` so a simple up when you need device approval
fixes the issue.

Updates #17382

Signed-off-by: Alex Chan <alexc@tailscale.com>