cmd/tailscale/cli: remember custom control server in `tailscale up`
This fixes an issue where a user logs in with a custom control server,
then later runs `tailscale up` without the `--login-server` argument,
and we "forget" their custom server.

This can manifest as asking them to log in again to login.tailscale.com,
or `tailscale up` completing successfully but the node staying down.
(If you look in the tailscaled logs, you see it requesting an auth URL
from login.tailscale.com, but not showing this to the user.)

The mistake is exiting `tailscale up` early if there are no flags,
without checking if the default flag values conflict with previously set
preferences.

The fix is to prevent the early return if the control URL changes, which
will trigger the subequent logic that compares cur/new prefs, which in
turn will flag the mismatch.

Fixes #17476

Signed-off-by: Alex Chan <alexc@tailscale.com>