tka/tailchonk: rewrite tests so they can be shared with corp
There's a TODO item in the tests for the SQLite-backed `dbchonk` in corp
that points out the oss/corp tests have a lot of copied code, which
could be reused.

This patch pulls the Chonk tests into a new `TestChonkImplementation`
helper which can be reused across repos. It also introduces a
`LastActiveAncestor` test which is in corp with a TODO "add to oss".

Updates tailscale/corp#27166

Signed-off-by: Alex Chan <alexc@tailscale.com>