tka: delete filesystem entries for purged AUMs
In clients, AUMs are stored as files on disk. Previously, when an AUM was purged, we'd do a soft-delete – we'd add metadata to mark the AUM as deleted,
but keep the underlying file.

This was a precautionary measure, so that we could recover purged AUMs
if we ever discovered a bug. After two years, we're confident this precaution
is no longer required.

This patch changes the behaviour, so now purged AUMs are deleted from the
local filesystem. This prevents unbounded growth of the AUM directory on
customer devices.

When an existing node first runs with this patch, any previously-purged
AUMs will temporarily "return", i.e. become available to the code.
The next time the node runs a compaction and purges old AUMs, they will
be re-purged and deleted from the disk.

The full AUM history is retained in the control plane, so we aren't deleting
any data permanently.

Signed-off-by: Alex Chan <alexc@tailscale.com>

Updates tailscale/corp#27166