tka: created a shared testing library for Chonk
This patch creates a set of tests that should be true for all
implementations of Chonk and CompactableChonk, which we can share
with the SQLite implementation in corp.

It includes all the existing tests, plus a test for LastActiveAncestor
which was in corp but not in oss.

Updates https://github.com/tailscale/corp/issues/33465

Signed-off-by: Alex Chan <alexc@tailscale.com>