util/eventbus: block for the subscriber during SubscribeFunc close
When a SubscribeFunc closes during the execution of a callback, wait for that
callback to finish before returning. We can't do anything about additional
goroutines the callback may have created (nor can we do the same for
Subscriber), but this ensures we treat callback return as the delivery action
rather than initiation. This should help avert shutdown races.

Updates #17638

Change-Id: I82fd31bcaa4e92fab07981ac0e57e6e3a7d9d60b
Signed-off-by: M. J. Fromberger <fromberger@tailscale.com>