wgengine: simplify how discokey change session dropping is implemented
When a discokey changes, we take that as a sign that the peer
restarted and our previous wireguard-go session state is invalid. What
we did previously was reconfig the wireguard-go Device with all peers
but the ones that changed, and then add them back.

Instead, just synchronously delete a single peer from wireguard-go
when we see that it's changed, and then let it be added back later by
the one Reconfig. This avoids a second reconfig and some twisty code.
(Plenty more twisty code yet to delete...)

Updates #17858

Change-Id: I084d51c26e368c909091da403bbf6b59d929e5af
Signed-off-by: Brad Fitzpatrick <bradfitz@tailscale.com>