tka: add a failing test for a tailchonk Compact issue
While working on #17942, I tried to construct and compact a chain of
AUMs, but I kept getting the error:

> reached genesis AUM without intersecting with candidate ancestor

This test reproduces the issue -- it's a simple chain `A -> B -> C` that
should be possible to compact, but it fails.

I don't have time to investigate why right now, but I'm putting this in
a branch to come back to later.

Change-Id: Ic2d303f0aaaad5d02ec79e9cc94e6a8953432bf0
Signed-off-by: Alex Chan <alexc@tailscale.com>