cmd/mkpkg: add changelog support for deb and rpm packages
Add a --changelog flag that accepts a path to a changelog.yaml file
in the format expected by goreleaser/chglog. When provided, the
changelog is automatically included in the package: deb packages
include it as /usr/share/doc/<package>/changelog.Debian.gz, and
rpm packages include it as changelog metadata in the package header.

The changelog flag is optional and only applies when a valid YAML
file path is provided.

Updates #314

Signed-off-by: Brendan Creane <bcreane@gmail.com>